% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.xtqsh}
\alias{coef.xtqsh}
\title{Extract Coefficients from xtqsh Objects}
\usage{
\method{coef}{xtqsh}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtqsh"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A matrix of coefficients with rows for each quantile and columns
  for each variable.
}
\description{
Returns the minimum distance quantile regression estimates.
}
