% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.xtqsh}
\alias{print.summary.xtqsh}
\title{Print Method for summary.xtqsh Objects}
\usage{
\method{print}{summary.xtqsh}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.xtqsh"}.}

\item{digits}{Number of significant digits. Default is 4.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for summary.xtqsh Objects
}
