% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.xtqsh}
\alias{print.xtqsh}
\title{Print Method for xtqsh Objects}
\usage{
\method{print}{xtqsh}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtqsh"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for xtqsh Objects
}
