% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.xtqsh}
\alias{summary.xtqsh}
\title{Summary Method for xtqsh Objects}
\usage{
\method{summary}{xtqsh}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtqsh"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
An object of class \code{"summary.xtqsh"}.
}
\description{
Summary Method for xtqsh Objects
}
